// migrations/snapshots/604af0e00d6abfc869d4103e15bcee6cbc19d1d5c8bbb50586ce4cb3c55bc26b/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly researchId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Conversation: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Finding: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly researchId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly data: CodecTypes['pg/json@1']['output'] | null;
      readonly displayStyle: 'TEXT';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly FindingSource: {
      readonly findingId: CodecTypes['pg/text@1']['output'];
      readonly sourceId: CodecTypes['pg/text@1']['output'];
    };
    readonly Membership: {
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly organizationId: CodecTypes['pg/text@1']['output'];
      readonly role: 'ADMIN' | 'MEMBER';
    };
    readonly Message: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly conversationId: CodecTypes['pg/text@1']['output'];
      readonly authorType: 'USER' | 'AI';
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Organization: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Research: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly workspaceId: CodecTypes['pg/text@1']['output'];
      readonly createdById: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'IN_PROGRESS' | 'COMPLETED' | 'ARCHIVED';
      readonly conclusion: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly ResearchTag: {
      readonly researchId: CodecTypes['pg/text@1']['output'];
      readonly tagId: CodecTypes['pg/text@1']['output'];
    };
    readonly Source: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly researchId: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly workspaceId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Workspace: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly organizationId: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly WorkspaceMembership: {
      readonly userId: CodecTypes['pg/text@1']['output'];
      readonly workspaceId: CodecTypes['pg/text@1']['output'];
      readonly role: 'ADMIN' | 'MEMBER';
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Comment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly researchId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Conversation: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Finding: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly researchId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly data: CodecTypes['pg/json@1']['input'] | null;
      readonly displayStyle: 'TEXT';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly FindingSource: {
      readonly findingId: CodecTypes['pg/text@1']['input'];
      readonly sourceId: CodecTypes['pg/text@1']['input'];
    };
    readonly Membership: {
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly organizationId: CodecTypes['pg/text@1']['input'];
      readonly role: 'ADMIN' | 'MEMBER';
    };
    readonly Message: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly conversationId: CodecTypes['pg/text@1']['input'];
      readonly authorType: 'USER' | 'AI';
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Organization: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Research: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly workspaceId: CodecTypes['pg/text@1']['input'];
      readonly createdById: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'IN_PROGRESS' | 'COMPLETED' | 'ARCHIVED';
      readonly conclusion: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly ResearchTag: {
      readonly researchId: CodecTypes['pg/text@1']['input'];
      readonly tagId: CodecTypes['pg/text@1']['input'];
    };
    readonly Source: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly researchId: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly workspaceId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Workspace: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly organizationId: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly WorkspaceMembership: {
      readonly userId: CodecTypes['pg/text@1']['input'];
      readonly workspaceId: CodecTypes['pg/text@1']['input'];
      readonly role: 'ADMIN' | 'MEMBER';
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;